- Init new CodeEditor component - Save the current index to draft - Register new filter for liquidJS engine
