+++ package.json
@@ -20,12 +20,16 @@
     "gatsby-plugin-netlify": "^2.0.6",
     "gatsby-plugin-react-helmet": "^3.0.4",
     "graphql-tag": "^2.10.1",
+    "liquidjs": "^10.7.0",
     "parcel-bundler": "^1.9.4",
+    "prismjs": "^1.29.0",
     "prop-types": "^15.6.0",
     "react": "^16.8.4",
     "react-apollo": "^2.5.5",
     "react-dom": "^16.8.4",
     "react-helmet": "^5.2.0",
+    "react-simple-code-editor": "^0.13.1",
+    "react-syntax-highlighter": "^15.5.0",
     "yarn-audit-fix": "^9.3.9"
   },
   "keywords": [
