+++ src/components/code/style.css
@@ -1,0 +1,3 @@
+.code-editor {
+    background-color: #fafafa;
+}
