+++ src/pages/index.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, {useState} from "react";
 import {
   Banner,
   Button,
@@ -20,21 +20,10 @@ import {
   Subheading,
 } from "@shopify/polaris";
 import faker from "faker";
-import gql from "graphql-tag";
-import { Mutation } from "react-apollo";
-import { parseGid } from "@shopify/admin-graphql-api-utilities";
+import { Liquid } from 'liquidjs';
 
 import PageLayout from "../components/layout";
-import SEO from "../components/seo";
-import GraphqlProvider from "../components/graphqlProvider";
-
-const hideIcon = (
-  <svg xmlns="http://www.w3.org/2000/svg" x="0px" y="0px" width="16px" height="16px" viewBox="0 0 16 16"><g transform="translate(0, 0)"><path fill="#637381" d="M14.574,5.669l-1.424,1.424C13.428,7.44,13.656,7.757,13.819,8c-0.76,1.13-2.85,3.82-5.561,3.985L6.443,13.8 C6.939,13.924,7.457,14,8,14c4.707,0,7.744-5.284,7.871-5.508c0.171-0.304,0.172-0.676,0.001-0.98 C15.825,7.427,15.372,6.631,14.574,5.669z"></path> <path fill="#637381" d="M0.293,15.707C0.488,15.902,0.744,16,1,16s0.512-0.098,0.707-0.293l14-14c0.391-0.391,0.391-1.023,0-1.414 s-1.023-0.391-1.414,0l-2.745,2.745C10.515,2.431,9.331,2,8,2C3.245,2,0.251,7.289,0.126,7.514 c-0.169,0.303-0.168,0.672,0.002,0.975c0.07,0.125,1.044,1.802,2.693,3.276l-2.529,2.529C-0.098,14.684-0.098,15.316,0.293,15.707z M2.181,7.999C2.958,6.835,5.146,4,8,4c0.742,0,1.437,0.201,2.078,0.508L8.512,6.074C8.348,6.029,8.178,6,8,6C6.895,6,6,6.895,6,8 c0,0.178,0.029,0.348,0.074,0.512L4.24,10.346C3.285,9.51,2.559,8.562,2.181,7.999z"></path></g></svg>
-)
-
-const viewIcon = (
-  <svg xmlns="http://www.w3.org/2000/svg" x="0px" y="0px" width="16px" height="16px" viewBox="0 0 16 16"><g transform="translate(0, 0)"><path fill="#637381" d="M8,14c4.707,0,7.744-5.284,7.871-5.508c0.171-0.304,0.172-0.676,0.001-0.98C15.746,7.287,12.731,2,8,2 C3.245,2,0.251,7.289,0.126,7.514c-0.169,0.303-0.168,0.672,0.002,0.975C0.254,8.713,3.269,14,8,14z M8,4 c2.839,0,5.036,2.835,5.818,4C13.034,9.166,10.837,12,8,12c-2.841,0-5.038-2.838-5.819-4.001C2.958,6.835,5.146,4,8,4z"></path> <circle data-color="color-2" fill="#637381" cx="8" cy="8" r="2"></circle></g></svg>
-)
+import Editor from "../components/code";
 
 const orderData = {
   email: faker.internet.email(),
@@ -63,148 +52,86 @@ const orderData = {
   tags: "gil-shopify-order-wizard",
 };
 
-const CREATE_DRAFT_ORDER = gql`
-  mutation draftOrderCreate($input: DraftOrderInput!) {
-    draftOrderCreate(input: $input) {
-      draftOrder {
-        id
+function IndexPage () {
+  const [liquidCode, setLiquidCode] = React.useState(
+    `{% assign name = 'John' %}
+    Hello, {{ name }}!`
+  );
+  const [parsedLiquidCode, setParsedLiquidCode] = React.useState('');
+
+  const engine = new Liquid();
+  // Register new filter to parse JSON string to object
+  engine.registerFilter("jsonToObject", (json) => { 
+    return JSON.parse(json)
+  });
+
+  const setMockData = (liquidCode) => {
+    const mockTemplate = `
+    {% capture mock_order %}
+    {
+        "attributes": {},
+        "billing_address": {},
+        "cancel_reason": null,
+        "cancel_reason_label": null,
+        "cancelled": false,
+        "cancelled_at": null,
+        "cart_level_discount_applications": [],
+        "created_at": "2022-04-29 11:15:46 -0400",
+        "customer": {},
+        "customer_order_url": "https://shopify.com/56174706753/account/orders/4295688749121?locale=en",
+        "customer_url": "https://polinas-potent-potions.myshopify.com/account/orders/8be02e56c658bcd1f034d28c496fddd9",
+        "discount_applications": [],
+        "discounts": null,
+        "email": "cornelius.potionmaker@gmail.com",
+        "financial_status": "paid",
+        "financial_status_label": "Paid",
+        "fulfillment_status": "partial",
+        "fulfillment_status_label": "Partial",
+        "id": 4295688749121,
+        "item_count": 6,
+        "line_items": [],
+        "line_items_subtotal_price": "492.93",
+        "metafields": {},
+        "name": "#1001",
+        "note": null,
+        "order_number": 1001,
+        "order_status_url": "https://polinas-potent-potions.myshopify.com/56174706753/orders/8be02e56c658bcd1f034d28c496fddd9/authenticate?key=4f9baf2b8ebd0f75ec73eb9bac6e4519",
+        "phone": null,
+        "shipping_address": {},
+        "shipping_methods": [],
+        "shipping_price": "0.00",
+        "subtotal_line_items": [],
+        "subtotal_price": "492.93",
+        "tags": [],
+        "tax_lines": [],
+        "tax_price": "0.00",
+        "total_discounts": "0.00",
+        "total_duties": null,
+        "total_net_amount": "492.93",
+        "total_price": "492.93",
+        "total_refunded_amount": "0.00",
+        "transactions": []
       }
-      userErrors {
-        field
-        message
-      }
+    {% endcapture %}
+    
+    {% assign order = mock_order | jsonToObject %}
+    `;
+
+    return `${mockTemplate}${liquidCode}`;
+  };
+
+  const handleCodeChange = (newCode) => {
+    setLiquidCode(newCode);
+    try {
+      const mockTemplate = setMockData(newCode);
+      const compiledTemplate = engine.parse(mockTemplate);
+      const renderedTemplate = engine.renderSync(compiledTemplate);
+      console.log('Rendered template:', renderedTemplate);
+      setParsedLiquidCode(renderedTemplate);
+    } catch (e) {
+      console.error('Error rendering template:', e);
     }
-  }
-`;
-
-const COMPLETE_DRAFT_ORDER = gql`
-  mutation draftOrderComplete($id: ID!) {
-    draftOrderComplete(id: $id) {
-      draftOrder {
-        order {
-          id
-        }
-      }
-      userErrors {
-        field
-        message
-      }
-    }
-  }
-`;
-
-class IndexPage extends React.Component {
-  state = {
-    token: null,
-    domain: null,
-    showStatusSection: false,
-    showProductSection: false,
-    showOrderModal: false,
-    isOrderSubmitting: false,
-    isDraftOrder: false,
-    isPasswordVisible: false,
-    order: orderData,
-    newOrder: null,
   };
-
-  componentDidMount() {
-    // TODO: retrieve domain cookie and set as state
-  }
-
-  handleDomainChange = domain => {
-    // TODO: save as cookie?
-    domain.replace("myshopify.com", "");
-    this.setState({ domain });
-  };
-
-  handleTokenChange = token => {
-    this.setState({ token });
-  };
-
-  handleTestOrderChange = () => {
-    this.setState({
-      order: {
-        ...this.state.order,
-        test: !this.state.order.test,
-      },
-    });
-  };
-
-  handleOrderTypeChange = () => {
-    this.setState({
-      isDraftOrder: !this.state.isDraftOrder,
-    });
-  };
-
-  handleOrderDataReset = () => {
-    this.setState({
-      order: orderData,
-    });
-  };
-
-  handleOrderNoteChange = note => {
-    this.setState({
-      order: {
-        ...this.state.order,
-        note,
-      },
-    });
-  };
-
-  handleOrderTagsChange = tags => {
-    this.setState({
-      order: {
-        ...this.state.order,
-        tags,
-      },
-    });
-  };
-
-  handlePasswordToggle = () => {
-    this.setState({
-      isPasswordVisible: !this.state.isPasswordVisible,
-    });
-  };
-
-  handleRegenerate = () => {
-    alert("TODO");
-  };
-
-  handleCreateOrder = () => {
-    this.handleOrderModal();
-  };
-
-  handleStatusSectionClick = () => {
-    this.setState({
-      showStatusSection: !this.state.showStatusSection,
-    });
-  };
-
-  handleProductSectionClick = () => {
-    this.setState({
-      showProductSection: !this.state.showProductSection,
-    });
-  };
-
-  handleOrderModal = () => {
-    this.setState({
-      showOrderModal: !this.state.showOrderModal,
-    });
-  };
-
-  render() {
-    const {
-      order,
-      domain,
-      token,
-      showOrderModal,
-      newOrder,
-      isDraftOrder,
-    } = this.state;
-    const privateAppUrl = domain
-      ? `//${domain}.myshopify.com/admin/apps/private/new`
-      : "https://help.shopify.com/en/manual/apps/private-apps";
-
     return (
       <PageLayout>
           <Page
@@ -226,386 +153,27 @@ class IndexPage extends React.Component {
                 </DisplayText>
               </Layout.Section>
               <Layout.Section>
-                <SEO
-                  title="Home"
-                  keywords={[`shopify`, `application`, `react`]}
-                />
-                {!token && (
-                  <Banner
-                    title="This site requires a private Shopify app. Enable the following permissions for the app:"
-                    action={{
-                      content: "Create Private App",
-                      url: privateAppUrl,
-                    }}
-                    status="info"
-                  >
-                    <TextContainer>
-                      <List type="bullet">
-                        <List.Item>
-                          Draft Orders - Read &amp; Write
-                        </List.Item>
-                        <List.Item>
-                          Orders, transactions and fulfillments - Read
-                          &amp; Write
-                        </List.Item>
-                      </List>
-                    </TextContainer>
-                  </Banner>
-                )}
-              </Layout.Section>
-              <Layout.Section>
-                <Card sectioned title="Settings">
+                <Card sectioned title="Liquid Editor">
                   <Stack distribution="fill">
-                    <TextField
-                      label="Shop Domain"
-                      value={domain}
-                      onChange={this.handleDomainChange}
-                      prefix="https://"
-                      suffix=".myshopify.com"
-                      placeholder="shopify-order-wizard"
-                    />
-                    <TextField
-                      type={
-                        this.state.isPasswordVisible
-                          ? "text"
-                          : "password"
-                      }
-                      label="Private App Password"
-                      value={token}
-                      onChange={this.handleTokenChange}
-                      helpText="We do not collect or store this information."
-                      connectedRight={
-                        <Button
-                          icon={this.state.isPasswordVisible ? hideIcon : viewIcon}
-                          label="Toggle Password"
-                          onClick={this.handlePasswordToggle}
-                        />
-                      }
-                    />
+                    <Editor code={liquidCode} onEdit={handleCodeChange}/>
+                  </Stack>
+                  <Stack distribution="fill">
+                    <Button>
+                      <Subheading>Copy Code</Subheading>
+                    </Button>
+                    <Button>
+                      <Subheading>Preview Code</Subheading>
+                    </Button>
+                  </Stack>
+                  <Stack distribution="fill">
+                  <Editor code={parsedLiquidCode} onEdit={()=>{}}/>
                   </Stack>
                 </Card>
-                <Card>
-                  <Card.Header title="Customer Information">
-                    {/* <Button
-                      icon="refresh"
-                      size="slim"
-                      onClick={this.handleRegenerate}
-                    >
-                      Regenerate
-                    </Button> */}
-                  </Card.Header>
-                  <Card.Section title="Customer">
-                    <TextContainer>
-                      <TextStyle variation="subdued">
-                        {order.shippingAddress.firstName}{" "}
-                        {order.shippingAddress.lastName}
-                      </TextStyle>
-                    </TextContainer>
-                    <TextContainer>
-                      <TextStyle variation="subdued">
-                        {order.email}
-                      </TextStyle>
-                    </TextContainer>
-                    <TextContainer>
-                      <TextStyle variation="subdued">
-                        {order.billingAddress.phone}
-                      </TextStyle>
-                    </TextContainer>
-                  </Card.Section>
-                  <Card.Section>
-                    <Stack distribution="fill">
-                      <>
-                        <Subheading>Shipping Address</Subheading>
-                        <br />
-                        <TextContainer>
-                          <TextStyle variation="subdued">
-                            {order.shippingAddress.firstName}{" "}
-                            {order.shippingAddress.lastName}
-                            <br />
-                            {order.shippingAddress.streetAddress}
-                            <br />
-                            {order.shippingAddress.city}{" "}
-                            {order.shippingAddress.province}{" "}
-                            {order.shippingAddress.zip}
-                            <br />
-                            {order.shippingAddress.country}
-                            <br />
-                            {order.shippingAddress.phone}
-                          </TextStyle>
-                        </TextContainer>
-                      </>
-                      <>
-                        <Subheading>Billing Address</Subheading>
-                        <br />
-                        <TextContainer>
-                          <TextStyle variation="subdued">
-                            {order.billingAddress.firstName}{" "}
-                            {order.billingAddress.lastName}
-                            <br />
-                            {order.billingAddress.streetAddress}
-                            <br />
-                            {order.billingAddress.city}{" "}
-                            {order.billingAddress.province}{" "}
-                            {order.billingAddress.zip}
-                            <br />
-                            {order.billingAddress.country}
-                            <br />
-                            {order.billingAddress.phone}
-                          </TextStyle>
-                        </TextContainer>
-                      </>
-                    </Stack>
-                  </Card.Section>
-                </Card>
-              {typeof window === "undefined" ? <></> :
-                <GraphqlProvider shop={domain} token={token}>
-                  <Mutation mutation={CREATE_DRAFT_ORDER}>
-                    {(draftOrderCreate, { data }) => (
-                      <Mutation mutation={COMPLETE_DRAFT_ORDER}>
-                        {(draftOrderComplete, { data }) => (
-                          <Card
-                            secondaryFooterAction={{
-                              content: "Reset",
-                              onAction: this.handleOrderDataReset,
-                            }}
-                            primaryFooterAction={{
-                              content: isDraftOrder
-                                ? "Create Draft Order"
-                                : "Create Order",
-                              loading: this.state.isOrderSubmitting,
-                              onAction: e => {
-                                e.preventDefault();
-
-                                this.setState({
-                                  isOrderSubmitting: true,
-                                });
-
-                                draftOrderCreate({
-                                  variables: {
-                                    input: {
-                                      lineItems: {
-                                        title: "Fries",
-                                        originalUnitPrice: "12",
-                                        quantity: 1,
-                                      },
-                                      ...this.state.order,
-                                      // appliedDiscount
-                                    },
-                                  },
-                                })
-                                  .then(res => {
-                                    const draftOrderId =
-                                      res &&
-                                      res.data &&
-                                      res.data.draftOrderCreate &&
-                                      res.data.draftOrderCreate.draftOrder &&
-                                      res.data.draftOrderCreate.draftOrder.id;
-
-                                    if (isDraftOrder) {
-                                      this.setState({
-                                        isOrderSubmitting: false,
-                                        showOrderModal: true,
-                                        newOrder: {
-                                          id: parseGid(draftOrderId),
-                                          type: "draft_orders",
-                                        },
-                                      });
-                                      return;
-                                    }
-
-                                    draftOrderComplete({
-                                      variables: {
-                                        id: draftOrderId,
-                                        paymentPending: true,
-                                      },
-                                    }).then(res => {
-                                      const orderId =
-                                        res &&
-                                        res.data &&
-                                        res.data.draftOrderComplete &&
-                                        res.data.draftOrderComplete.draftOrder &&
-                                        res.data.draftOrderComplete.draftOrder
-                                          .order &&
-                                        res.data.draftOrderComplete.draftOrder
-                                          .order.id;
-
-                                      this.setState({
-                                        isOrderSubmitting: false,
-                                        showOrderModal: true,
-                                        newOrder: {
-                                          id: parseGid(orderId),
-                                          type: "orders",
-                                        },
-                                      });
-                                    });
-                                  })
-                                  .catch(e => {
-                                    // TODO: show error message in modal or somewhere...
-                                    this.setState({
-                                      isOrderSubmitting: false,
-                                    });
-                                  });
-                              },
-                              disabled: !domain || !token,
-                            }}
-                          >
-                            <Card.Header title="Order Info">
-                              {/* <Checkbox
-                            checked={order.test}
-                            label="Test Order"
-                            onChange={this.handleTestOrderChange}
-                          /> */}
-                              <Checkbox
-                                checked={isDraftOrder}
-                                label="Draft Order"
-                                onChange={this.handleOrderTypeChange}
-                              />
-                            </Card.Header>
-                            <Card.Section>
-                              <FormLayout>
-                                <TextField
-                                  label="Tags"
-                                  onChange={this.handleOrderTagsChange}
-                                  value={order.tags}
-                                  helpText="Comma Separated"
-                                />
-                                {/* <TextField
-                            label="Source Name"
-                            onChange={this.handleChange}
-                            value="gil_greenberg_order_wizard"
-                            disabled
-                          /> */}
-                                <TextField
-                                  label="Order Note"
-                                  value={order.note}
-                                  onChange={this.handleOrderNoteChange}
-                                  maxLength={120}
-                                  showCharacterCount
-                                  multiline
-                                />
-                              </FormLayout>
-                            </Card.Section>
-                            {/* <Card.Section>
-                              <Button
-                                icon={
-                                  this.state.showStatusSection
-                                    ? "subtract"
-                                    : "add"
-                                }
-                                plain
-                                onClick={this.handleStatusSectionClick}
-                                ariaExpanded={this.state.showStatusSection}
-                              >
-                                <Subheading>Order Statuses</Subheading>
-                              </Button>
-                              <Collapsible
-                                open={this.state.showStatusSection}
-                                id="status-collapsible"
-                              >
-                                <br />
-                                <FormLayout>
-                                  <Select
-                                    label="Financial Status"
-                                    options={[
-                                      {
-                                        label: "authorized",
-                                        value: "authorized",
-                                      },
-                                      { label: "pending", value: "pending" },
-                                      { label: "paid", value: "paid" },
-                                      { label: "refunded", value: "refunded" },
-                                      { label: "voided", value: "voided" },
-                                    ]}
-                                    onChange={this.handleChange}
-                                    value={this.state.selected}
-                                  />
-                                  <Select
-                                    label="Fulfillment Status"
-                                    options={[
-                                      { label: "unshipped", value: "unshipped" },
-                                      { label: "partial", value: "partial" },
-                                      { label: "shipped", value: "shipped" },
-                                    ]}
-                                    onChange={this.handleChange}
-                                    value={this.state.selected}
-                                  />
-                                  <Select
-                                    label="Order Status"
-                                    options={[
-                                      { label: "Open", value: "open" },
-                                      { label: "Closed", value: "closed" },
-                                    ]}
-                                    onChange={this.handleChange}
-                                    value={this.state.selected}
-                                  />
-                                </FormLayout>
-                              </Collapsible>
-                            </Card.Section>
-                            <Card.Section>
-                              <Button
-                                icon={
-                                  this.state.showProductSection
-                                    ? "subtract"
-                                    : "add"
-                                }
-                                plain
-                                onClick={this.handleProductSectionClick}
-                                ariaExpanded={this.state.showProductSection}
-                              >
-                                <Subheading>Product Data</Subheading>
-                              </Button>
-                              <Collapsible
-                                open={this.state.showProductSection}
-                                id="product-collapsible"
-                              >
-                                <br />
-                                <FormLayout>
-                                  <Heading>
-                                    Custom Product Checkbox otherwise variant
-                                  </Heading>
-                                  Variant ID Product title qty price has shipping
-                                  <Heading>Order Discount</Heading>
-                                  <Heading>Order Risk</Heading>
-                                  TODO:
-                                  https://help.shopify.com/en/api/reference/orders/order-risk#create
-                                </FormLayout>
-                              </Collapsible>
-                            </Card.Section> */}
-                            <Card.Section />
-                          </Card>
-                        )}
-                      </Mutation>
-                    )}
-                  </Mutation>
-                </GraphqlProvider>
-              }
               </Layout.Section>
             </Layout>
-            <Modal
-              open={showOrderModal}
-              onClose={this.handleOrderModal}
-              sectioned
-            >
-              <Modal.Section>
-                <TextContainer>
-                  <Heading>Order Successfully Created!</Heading>
-                  <Button
-                    primary
-                    external
-                    icon="external"
-                    url={`//${domain}.myshopify.com/admin/${(newOrder &&
-                      newOrder.type) ||
-                      "orders"}/${(newOrder && newOrder.id) || ""}`}
-                  >
-                    View Order
-                  </Button>
-                </TextContainer>
-              </Modal.Section>
-            </Modal>
           </Page>
         </PageLayout>
     );
-  }
 }
 
 export default IndexPage;
